feat(api): create combined-course-details-service
