+++ api/src/quest/domain/models/CombinedCourse.js
@@ -58,6 +58,7 @@ export class CombinedCourseDetails extends CombinedCourse {
           ? CombinedCourseStatuses.STARTED
           : CombinedCourseStatuses.COMPLETED;
     }
+    this.participation = participation;
   }
 
   get campaignIds() {
