+++ api/tests/quest/integration/domain/usecases/get-combined-course-by-code_test.js
@@ -136,253 +136,5 @@ describe('Integration | Quest | Domain | UseCases | get-combined-course-by-code'
       expect(result.items[1]).instanceOf(CombinedCourseItem);
       expect(result.items[2]).instanceOf(CombinedCourseItem);
     });
-
-    it('should return started combined course for given userId', async function () {
-      nock('https://assets.pix.org').persist().head(/^.+$/).reply(200, {});
-      const { id: organizationLearnerId, userId, organizationId } = databaseBuilder.factory.buildOrganizationLearner();
-      const targetProfile = databaseBuilder.factory.buildTargetProfile({ ownerOrganizationId: organizationId });
-      const campaign = databaseBuilder.factory.buildCampaign({ targetProfileId: targetProfile.id, organizationId });
-      const campaignParticipation = databaseBuilder.factory.buildCampaignParticipation({
-        campaignId: campaign.id,
-        userId,
-        organizationLearnerId,
-        status: 'SHARED',
-      });
-      const training1 = databaseBuilder.factory.buildTraining({ type: 'modulix', link: '/modules/bac-a-sable' });
-      const training2 = databaseBuilder.factory.buildTraining({ type: 'modulix', link: '/modules/bases-clavier-1' });
-      databaseBuilder.factory.buildTraining({
-        type: 'modulix',
-        link: '/modules/bien-ecrire-son-adresse-mail',
-      });
-      const moduleId1 = '6282925d-4775-4bca-b513-4c3009ec5886';
-      const moduleId2 = '654c44dc-0560-4acc-9860-4a67c923577f';
-      const moduleId3 = 'f7b3a2e1-0d5c-4c6c-9c4d-1a3d8f7e9f5d';
-      databaseBuilder.factory.buildTargetProfileTraining({
-        targetProfileId: targetProfile.id,
-        trainingId: training1.id,
-      });
-      databaseBuilder.factory.buildTargetProfileTraining({
-        targetProfileId: targetProfile.id,
-        trainingId: training2.id,
-      });
-      databaseBuilder.factory.buildUserRecommendedTraining({
-        userId,
-        trainingId: training1.id,
-        campaignParticipationId: campaignParticipation.id,
-      });
-      databaseBuilder.factory.buildPassage({ moduleId: moduleId1, userId, terminatedAt: new Date() });
-
-      const { id: questId } = databaseBuilder.factory.buildQuestForCombinedCourse({
-        code,
-        organizationId,
-        successRequirements: [
-          {
-            requirement_type: 'campaignParticipations',
-            comparison: 'all',
-            data: {
-              campaignId: {
-                data: campaign.id,
-                comparison: 'equal',
-              },
-              status: {
-                data: 'SHARED',
-                comparison: 'equal',
-              },
-            },
-          },
-          {
-            requirement_type: 'passages',
-            comparison: 'all',
-            data: {
-              moduleId: {
-                data: moduleId1,
-                comparison: 'equal',
-              },
-              isTerminated: {
-                data: true,
-                comparison: 'equal',
-              },
-            },
-          },
-          {
-            requirement_type: 'passages',
-            comparison: 'all',
-            data: {
-              moduleId: {
-                data: moduleId2,
-                comparison: 'equal',
-              },
-              isTerminated: {
-                data: true,
-                comparison: 'equal',
-              },
-            },
-          },
-          {
-            requirement_type: 'passages',
-            comparison: 'all',
-            data: {
-              moduleId: {
-                data: moduleId3,
-                comparison: 'equal',
-              },
-              isTerminated: {
-                data: true,
-                comparison: 'equal',
-              },
-            },
-          },
-        ],
-      });
-
-      databaseBuilder.factory.buildCombinedCourseParticipation({ questId, organizationLearnerId });
-
-      await databaseBuilder.commit();
-
-      const result = await usecases.getCombinedCourseByCode({ code, userId });
-
-      expect(result.items).to.be.deep.equal([
-        {
-          id: campaign.id,
-          reference: campaign.code,
-          title: campaign.title,
-          type: ITEM_TYPE.CAMPAIGN,
-          redirection: undefined,
-          isCompleted: true,
-          isLocked: false,
-          duration: undefined,
-          image: undefined,
-        },
-        {
-          id: moduleId1,
-          reference: 'bac-a-sable',
-          title: 'Bac à sable',
-          type: ITEM_TYPE.MODULE,
-          redirection: 'encryptedUrl',
-          isCompleted: true,
-          isLocked: false,
-          duration: 5,
-          image: 'https://assets.pix.org/modules/placeholder-details.svg',
-        },
-        {
-          id: moduleId3,
-          reference: 'bien-ecrire-son-adresse-mail',
-          title: 'Bien écrire une adresse mail',
-          type: ITEM_TYPE.MODULE,
-          redirection: 'encryptedUrl',
-          isCompleted: false,
-          isLocked: false,
-          duration: 10,
-          image: 'https://assets.pix.org/modules/bien-ecrire-son-adresse-mail-details.svg',
-        },
-      ]);
-      expect(result).to.be.instanceOf(CombinedCourse);
-      expect(result.id).to.equal(questId);
-      expect(result.status).to.equal(CombinedCourseStatuses.STARTED);
-      expect(result.items[0]).instanceOf(CombinedCourseItem);
-      expect(result.items[1]).instanceOf(CombinedCourseItem);
-      expect(result.items[2]).instanceOf(CombinedCourseItem);
-    });
-  });
-  describe('when there is no organization learner yet', function () {
-    it('should return correct data for a not started combined course participation', async function () {
-      nock('https://assets.pix.org').persist().head(/^.+$/).reply(200, {});
-
-      const organizationId = databaseBuilder.factory.buildOrganization().id;
-      const targetProfile = databaseBuilder.factory.buildTargetProfile({ ownerOrganizationId: organizationId });
-      const campaign = databaseBuilder.factory.buildCampaign({ targetProfileId: targetProfile.id, organizationId });
-      const userId = databaseBuilder.factory.buildUser().id;
-
-      const training1 = databaseBuilder.factory.buildTraining({ type: 'modulix', link: '/modules/bac-a-sable' });
-      const training2 = databaseBuilder.factory.buildTraining({ type: 'modulix', link: '/modules/bases-clavier-1' });
-      databaseBuilder.factory.buildTraining({
-        type: 'modulix',
-        link: '/modules/bien-ecrire-son-adresse-mail',
-      });
-      const moduleId1 = '6282925d-4775-4bca-b513-4c3009ec5886';
-      const moduleId2 = '654c44dc-0560-4acc-9860-4a67c923577f';
-      const moduleId3 = 'f7b3a2e1-0d5c-4c6c-9c4d-1a3d8f7e9f5d';
-      databaseBuilder.factory.buildTargetProfileTraining({
-        targetProfileId: targetProfile.id,
-        trainingId: training1.id,
-      });
-      databaseBuilder.factory.buildTargetProfileTraining({
-        targetProfileId: targetProfile.id,
-        trainingId: training2.id,
-      });
-
-      const { id: questId } = databaseBuilder.factory.buildQuestForCombinedCourse({
-        code,
-        organizationId,
-        successRequirements: [
-          {
-            requirement_type: 'campaignParticipations',
-            comparison: 'all',
-            data: {
-              campaignId: {
-                data: campaign.id,
-                comparison: 'equal',
-              },
-              status: {
-                data: 'SHARED',
-                comparison: 'equal',
-              },
-            },
-          },
-          {
-            requirement_type: 'passages',
-            comparison: 'all',
-            data: {
-              moduleId: {
-                data: moduleId1,
-                comparison: 'equal',
-              },
-              isTerminated: {
-                data: true,
-                comparison: 'equal',
-              },
-            },
-          },
-          {
-            requirement_type: 'passages',
-            comparison: 'all',
-            data: {
-              moduleId: {
-                data: moduleId2,
-                comparison: 'equal',
-              },
-              isTerminated: {
-                data: true,
-                comparison: 'equal',
-              },
-            },
-          },
-          {
-            requirement_type: 'passages',
-            comparison: 'all',
-            data: {
-              moduleId: {
-                data: moduleId3,
-                comparison: 'equal',
-              },
-              isTerminated: {
-                data: true,
-                comparison: 'equal',
-              },
-            },
-          },
-        ],
-      });
-
-      await databaseBuilder.commit();
-
-      const result = await usecases.getCombinedCourseByCode({ code, userId });
-      expect(result).to.be.instanceOf(CombinedCourse);
-      expect(result.id).to.equal(questId);
-      expect(result.status).to.equal(CombinedCourseStatuses.NOT_STARTED);
-      expect(result.items[0]).instanceOf(CombinedCourseItem);
-      expect(result.items[1]).instanceOf(CombinedCourseItem);
-      expect(result.items[2]).instanceOf(CombinedCourseItem);
-    });
   });
 });
