+++ api/tests/quest/unit/domain/services/combined-course-details-service_test.js
@@ -1,4 +1,4 @@
-import { getCombinedCourseByCode } from '../../../../../src/quest/domain/usecases/get-combined-course-by-code.js';
+import CombinedCourseDetailsService from '../../../../../src/quest/domain/services/combined-course-details-service.js';
 import { catchErr, expect, sinon } from '../../../../test-helper.js';
 
 describe('Quest | Unit | Domain | Usecases | getCombinedCourseByCode', function () {
@@ -6,19 +6,19 @@ describe('Quest | Unit | Domain | Usecases | getCombinedCourseByCode', function 
     // given
     const userId = Symbol('userId');
     const code = Symbol('code');
-    const questId = Symbol('questId');
     const combinedCourse = Symbol('CombinedCourse');
+    const quest = Symbol('Quest');
 
-    const questRepositoryStub = { getByCode: sinon.stub() };
-    const combinedCourseRepositoryStub = { getByCode: sinon.stub() };
+    const questRepositoryStub = { findById: sinon.stub() };
+    const combinedCourseRepositoryStub = { getById: sinon.stub() };
     const combinedCourseParticipationRepositoryStub = { getByUserId: sinon.stub() };
 
-    questRepositoryStub.getByCode.resolves({ questId, userId });
-    combinedCourseRepositoryStub.getByCode.resolves(combinedCourse);
+    questRepositoryStub.findById.resolves(quest);
+    combinedCourseRepositoryStub.getById.resolves(combinedCourse);
     combinedCourseParticipationRepositoryStub.getByUserId.rejects(new Error('oops'));
 
     // when
-    const error = await catchErr(getCombinedCourseByCode)({
+    const error = await catchErr(CombinedCourseDetailsService.getCombinedCourseDetails)({
       userId,
       code,
       questRepository: questRepositoryStub,
